La til egne klasse for validering av barnets vilkår
