+++ src/test/enhetstester/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vilkårsvurdering/VilkårsvurderingStegTest.kt
@@ -1,14 +1,15 @@
 package no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering
 
+import org.hamcrest.CoreMatchers.`is` as Is
 import io.mockk.every
-import io.mockk.impl.annotations.InjectMockKs
-import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.mockkObject
 import io.mockk.runs
 import io.mockk.verify
+import java.math.BigDecimal
+import java.time.LocalDate
 import no.nav.familie.ks.sak.api.dto.BarnMedOpplysningerDto
 import no.nav.familie.ks.sak.api.dto.SøkerMedOpplysningerDto
 import no.nav.familie.ks.sak.api.dto.SøknadDto
@@ -36,6 +37,11 @@ import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Utd
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkår
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.VilkårResultat
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkårsvurdering
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.validering.BarnetsAlderVilkårValidator
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.validering.BarnetsAlderVilkårValidator2021
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.validering.BarnetsAlderVilkårValidator2021og2024
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.validering.BarnetsAlderVilkårValidator2024
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.validering.BarnetsVilkårValidator
 import no.nav.familie.ks.sak.kjerne.beregning.BeregningService
 import no.nav.familie.ks.sak.kjerne.eøs.kompetanse.KompetanseService
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.PersonopplysningGrunnlagService
@@ -45,32 +51,39 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.api.extension.ExtendWith
-import java.math.BigDecimal
-import java.time.LocalDate
-import org.hamcrest.CoreMatchers.`is` as Is
 
 @ExtendWith(MockKExtension::class)
 class VilkårsvurderingStegTest {
-    @MockK
-    private lateinit var personopplysningGrunnlagService: PersonopplysningGrunnlagService
+    private val personopplysningGrunnlagService: PersonopplysningGrunnlagService = mockk()
+    private val behandlingService: BehandlingService = mockk()
+    private val vilkårsvurderingService: VilkårsvurderingService = mockk()
+    private val søknadGrunnlagService: SøknadGrunnlagService = mockk()
+    private val beregningService: BeregningService = mockk()
+    private val kompetanseService: KompetanseService = mockk()
 
-    @MockK
-    private lateinit var behandlingService: BehandlingService
-
-    @MockK
-    private lateinit var vilkårsvurderingService: VilkårsvurderingService
-
-    @MockK
-    private lateinit var søknadGrunnlagService: SøknadGrunnlagService
-
-    @MockK
-    private lateinit var beregningService: BeregningService
-
-    @MockK
-    private lateinit var kompetanseService: KompetanseService
-
-    @InjectMockKs
-    private lateinit var vilkårsvurderingSteg: VilkårsvurderingSteg
+    private val barnetsAlderVilkårValidator2021 = BarnetsAlderVilkårValidator2021()
+    private val barnetsAlderVilkårValidator2024 = BarnetsAlderVilkårValidator2024()
+    private val barnetsVilkårValidator: BarnetsVilkårValidator =
+        BarnetsVilkårValidator(
+            BarnetsAlderVilkårValidator(
+                barnetsAlderVilkårValidator2021,
+                barnetsAlderVilkårValidator2024,
+                BarnetsAlderVilkårValidator2021og2024(
+                    barnetsAlderVilkårValidator2021,
+                    barnetsAlderVilkårValidator2024,
+                ),
+            ),
+        )
+    private val vilkårsvurderingSteg: VilkårsvurderingSteg =
+        VilkårsvurderingSteg(
+            personopplysningGrunnlagService,
+            behandlingService,
+            søknadGrunnlagService,
+            vilkårsvurderingService,
+            beregningService,
+            kompetanseService,
+            barnetsVilkårValidator,
+        )
 
     private val søker = randomAktør()
     private val barn = randomAktør()
