+++ src/test/enhetstester/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vilkårsvurdering/validering/BarnetsAlderVilkårValidator2021og2024Test.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.validering
+
+import org.junit.jupiter.api.Test
+
+class BarnetsAlderVilkårValidator2021og2024Test {
+    @Test
+    fun `asdf`() {
+    }
+}
